+++ CHANGELOG.md
@@ -60,6 +60,7 @@ __New Feature__:
 - single command setup and run using starlake.sh / starlake.cmd
 - Updated quickstart with docker use
 - Infer schema now recognize date as date not timestamp
+- Support metadata folder comparison and produce HTML report
 
 # 0.5.2
 __New Feature__: